requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
mlxtend
plotly
fpdf
Pillow
sqlalchemy
bcrypt
email-validator
